ネットワーク生成と分析のための新しいクラスを追加し、古いネットワークモデルを削除。__init__.pyを更新して新しいクラスをインポート。#3
